requirements.txt
requests>=2.31.0
PyMuPDF>=1.23.8
Pillow>=10.0.0
pytesseract>=0.3.10
tiktoken>=0.5.2
tqdm>=4.66.0
supabase>=2.0.0
python-dotenv>=1.0.0
openai>=1.12.0
anthropic>=0.18.0
numpy>=1.24.0
google-generativeai>=0.3.0
cohere>=5.0.0
huggingface-hub>=0.20.0
flask>=3.0.0
